Server/requirements.txt
# Chanakya Server Dependencies
# ============================

# Core
pydantic>=2.0.0
python-dateutil>=2.8.0

# Gemini AI (new package)
google-genai>=1.0.0

# Optional: IndicBERT (for future fine-tuning)
# transformers>=4.35.0
# torch>=2.0.0
# sentencepiece>=0.1.99

# Async support
aiohttp>=3.9.0

# Logging and monitoring
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
python-dotenv>=1.0.0
